+++ NEWS.md
@@ -1,10 +1,14 @@
-## [geoflow 1.2.1.9000](https://github.com/r-geoflow/geoflow) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-unavailable-orange.svg)](https://github.com/r-geoflow/geoflow)
+## [geoflow 1.2.9000](https://github.com/r-geoflow/geoflow) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-unavailable-orange.svg)](https://github.com/r-geoflow/geoflow)
 
 
 **Bug fixes**
 
 * [#443](https://github.com/r-geoflow/geoflow/issues/443) Missing sourceSql export in entity$asDataFrame
 * [#444](https://github.com/r-geoflow/geoflow/issues/444) Data dir is ignored with entity$asDataFrame
+
+**New features**
+
+* [#445](https://github.com/r-geoflow/geoflow/issues/445) Support specific software configuration for environment file remote fetching
 
 ## [geoflow 1.2.1](https://github.com/r-geoflow/geoflow) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://github.com/r-geoflow/geoflow)